task: directory restructure
