+++ dataworkspace/dataworkspace/static/js/react/features/your-files/icons/TrashIcon.js
