+++ dataworkspace/dataworkspace/static/js/react/features/your-files/styles/App.scss
